changed some data type to optimise ummm... 8 temporary bytes ? lol

+++ CMinecraftGL/src/renderer/gl/vao.cpp
@@ -21,7 +21,8 @@ void VertexArray::unbind() const {
 	glCall (glBindVertexArray(0));
 }
 
-void VertexArray::link_attrib(VertexBuffer *vbo, GLuint va_index, GLint data_dim, GLenum type, GLsizei stride, int offset) {
+void VertexArray::link_attrib(VertexBuffer *vbo, GLuint va_index, 
+	unsigned short data_dim, GLenum type, GLsizei stride, int offset) {
 	bind();
 	vbo->bind();
 	glCall (glVertexAttribPointer(va_index, data_dim, GL_FLOAT, GL_FALSE, stride, (const void*)offset));
